@@ -19,12 +19,12 @@ version: 0.11.0
 homepage: https://github.com/sarahec/instant_xml_parser
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  logging: ^0.11.4
-  xml: ^4.4.0
-  ixp_runtime: ^0.11.0
+  logging: ^1.0.0
+  xml: ^5.0.2
+  ixp_runtime: ^1.0.0-nullsafety.0
 
 dev_dependencies:
   build_runner: ^1.0.0
@@ -38,3 +38,5 @@ dependency_overrides:
     path: ../instant_xml_parser
   ixp_runtime:
     path: ../ixp_runtime
+  logging: ^1.0.0
+  xml: ^5.0.2
